+++ Dyna/ExternalAPIs/IMoreDifficultiesApi.cs
@@ -1,6 +1,10 @@
-﻿namespace Shockah.Dyna;
+﻿using System;
+
+namespace Shockah.Dyna;
 
 public interface IMoreDifficultiesApi
 {
 	void RegisterAltStarters(Deck deck, StarterDeck starterDeck);
+
+	Type BasicOffencesCardType { get; }
 }